+++ packages/layout/src/demos/classicMode.tsx
@@ -5,9 +5,9 @@ import {
   QuestionCircleFilled,
   SearchOutlined,
 } from '@ant-design/icons';
-import ProCard from '@ant-design/pro-card';
-import type { ProSettings } from '@ant-design/pro-layout';
-import { PageContainer, ProLayout, SettingDrawer } from '@ant-design/pro-layout';
+import { ProCard } from '@ant-design/pro-components';
+import type { ProSettings } from '@ant-design/pro-components';
+import { PageContainer, ProLayout, SettingDrawer } from '@ant-design/pro-components';
 import { Button, Input } from 'antd';
 import { useState } from 'react';
 import defaultProps from './_defaultProps';