+++ packages/layout/src/demos/footer-global-tools.tsx
@@ -5,9 +5,9 @@ import {
   QuestionCircleFilled,
   SearchOutlined,
 } from '@ant-design/icons';
-import ProCard from '@ant-design/pro-card';
-import type { ProSettings } from '@ant-design/pro-layout';
-import { PageContainer, ProLayout } from '@ant-design/pro-layout';
+import { ProCard } from '@ant-design/pro-components';
+import type { ProSettings } from '@ant-design/pro-components';
+import { PageContainer, ProLayout } from '@ant-design/pro-components';
 import { Alert, Button, Input, Space } from 'antd';
 import React, { useState } from 'react';
 import defaultProps from './_defaultProps';