+++ tests/layout/mobile.test.tsx
@@ -1,4 +1,4 @@
-﻿import { ProLayout } from '@ant-design/pro-layout';
+﻿import { ProLayout } from '@ant-design/pro-components';
 import { render as reactRender, render } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
 import { waitForComponentToPaint } from '../util';