+++ packages/form/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.6...@ant-design/pro-form@2.4.7) (2022-12-13)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [2.4.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.5...@ant-design/pro-form@2.4.6) (2022-12-09)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [2.4.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.4...@ant-design/pro-form@2.4.5) (2022-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [2.4.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.3...@ant-design/pro-form@2.4.4) (2022-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [2.4.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.2...@ant-design/pro-form@2.4.3) (2022-12-06)
+
+### Bug Fixes
+
+- **form:** fix style error ([3320a52](https://github.com/ant-design/pro-components/commit/3320a528131b4cb49f70d76e9d5d6c6ec8ca510f))
+
+## [2.4.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.1...@ant-design/pro-form@2.4.2) (2022-12-05)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
 ## [2.4.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.4.0...@ant-design/pro-form@2.4.1) (2022-12-01)
 
 **Note:** Version bump only for package @ant-design/pro-form