+++ packages/provider/CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.5...@ant-design/pro-provider@2.2.0) (2022-12-13)
+
+### Features
+
+- **components:** bump swr from 1.x to 2.x ([#6367](https://github.com/ant-design/pro-components/issues/6367)) ([d7a5206](https://github.com/ant-design/pro-components/commit/d7a52068b9c70d40ecf68a065ed470573daade72))
+- **provide:** 繁體中文補充 ([#6358](https://github.com/ant-design/pro-components/issues/6358)) ([aa87579](https://github.com/ant-design/pro-components/commit/aa87579748b8a9e666c626ff10d94378336e296d))
+
+## [2.1.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.4...@ant-design/pro-provider@2.1.5) (2022-12-09)
+
+### Bug Fixes
+
+- export usestyle ([2834f23](https://github.com/ant-design/pro-components/commit/2834f23517116d2f2276cce403b5ad03662d9bf1))
+
+## [2.1.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.3...@ant-design/pro-provider@2.1.4) (2022-12-08)
+
+### Bug Fixes
+
+- **layout:** remove antdTheme ([9d0e057](https://github.com/ant-design/pro-components/commit/9d0e057fba7ea7fa546e4ad914106a14c4b1b244))
+
+## [2.1.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.2...@ant-design/pro-provider@2.1.3) (2022-12-08)
+
+### Bug Fixes
+
+- **layout:** support configprovide darkAlgorithm ([dabf05e](https://github.com/ant-design/pro-components/commit/dabf05edc5f0ef3edd58f69ecdab43981f02e044))
+
+## [2.1.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.1...@ant-design/pro-provider@2.1.2) (2022-12-05)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
 ## [2.1.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.1.0...@ant-design/pro-provider@2.1.1) (2022-12-01)
 
 ### Bug Fixes