+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.5...@ant-design/pro-utils@2.5.0) (2022-12-13)
+
+### Features
+
+- **components:** bump swr from 1.x to 2.x ([#6367](https://github.com/ant-design/pro-components/issues/6367)) ([d7a5206](https://github.com/ant-design/pro-components/commit/d7a52068b9c70d40ecf68a065ed470573daade72))
+
+## [2.4.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.4...@ant-design/pro-utils@2.4.5) (2022-12-09)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
+## [2.4.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.3...@ant-design/pro-utils@2.4.4) (2022-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
+## [2.4.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.2...@ant-design/pro-utils@2.4.3) (2022-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
+## [2.4.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.1...@ant-design/pro-utils@2.4.2) (2022-12-05)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
 ## [2.4.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.4.0...@ant-design/pro-utils@2.4.1) (2022-12-01)
 
 **Note:** Version bump only for package @ant-design/pro-utils